refactoring index actions seperate files.
